Fix CI formatting and stale renderer audit expectations after rebasing

+++ crates/gpui_macos/src/window.rs
@@ -1774,7 +1774,7 @@ impl PlatformWindow for MacWindow {
             button.setTag(ix as NSInteger);
 
             if answer.is_cancel() {
-                if let Some(key) = core::char::from_u32(crate::events::ESCAPE_KEY as u32) {
+                if let Some(key) = core::char::from_u32(crate::events::ESCAPE_KEY) {
                     let key = NSString::from_str(&key.to_string());
                     button.setKeyEquivalent(&key);
                 }
